Add another overloaded method for checking negative values

+++ Review/src/main/java/org/example/BankAccount.java
@@ -1,19 +1,22 @@
 package org.example;
 
+import java.math.BigDecimal;
+
 public class BankAccount {
 
     private static String bankName;
     private final int accountNumber;
-    private final double balance;
+    private final BigDecimal balance;
 
-    public BankAccount(String bankName, int accountNumber, double balance) {
+    public BankAccount(String bankName, int accountNumber, BigDecimal balance) {
         BankAccount.bankName = bankName;
         this.accountNumber =
-            ValueValidator.checkForNegativeValue(accountNumber, "Account number is less than zero");
-        this.balance = ValueValidator.checkForNegativeValue(balance, "Balance is less than zero");
+                ValueValidator.checkForNegativeValue(accountNumber, "Account number is less than zero");
+        this.balance =
+                ValueValidator.checkForNegativeValue(balance, "Balance is less than zero");
     }
 
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 }
